[refactor:move] app-api: tests/mod.rs のヘルパを tests/support/ へ分割
1,167行の tests/mod.rs から共有ヘルパ40アイテム(fn 32 + struct 8)を
support.rs(ハブ)+ support/{timeouts,waiters,doubles,iroh_stack,fixtures}.rs
へ移動。mod.rs は use 群 + テーマ mod 宣言 + シムのみの37行に縮小。
モジュール名は extern crate iroh との E0659 衝突を避け iroh_stack とした。

既存テーマ24ファイル(139テスト)は無変更。純粋な行移動からの唯一の
逸脱は pub(crate) の付与(アイテム本体・テストから直接参照される
struct フィールド10箇所・inherent impl メソッド9本。crate 外公開は不変)。

検証: nextest list 139テスト完全一致、rust-test green、
clippy -D warnings / fmt green、oversized-files 解消(最大322行)。

Refs: .claude/plans/2026-07-03-refactor-s1-test-split.md (T3 / 付録C)
